viewer prototype

+++ src/adapters/WebSQLAdapter.js
@@ -1,12 +1,14 @@
 export default class WebSQLAdapter {
-    constructor() {
+    constructor(create=true) {
         let db = openDatabase('vkdb', '1.0', 'Database with messages exported from VK', 2 * 1024 * 1024,() => {})
-        db.transaction(tx => {
-            tx.executeSql("DROP TABLE IF EXISTS users",[])
-            tx.executeSql("DROP TABLE IF EXISTS messages",[])
-            tx.executeSql("CREATE TABLE users (id INTEGER PRIMARY KEY, name TEXT)")
-            tx.executeSql("CREATE TABLE messages (id INTEGER PRIMARY KEY, uid INTEGER, cid INTEGER, txt TEXT, att TEXT, date INTEGER, FOREIGN KEY (uid) REFERENCES users(id))")
-        })
+        if (create) {
+            db.transaction(tx => {
+                tx.executeSql("DROP TABLE IF EXISTS users",[])
+                tx.executeSql("DROP TABLE IF EXISTS messages",[])
+                tx.executeSql("CREATE TABLE users (id INTEGER PRIMARY KEY, name TEXT)")
+                tx.executeSql("CREATE TABLE messages (id INTEGER PRIMARY KEY, uid INTEGER, cid INTEGER, txt TEXT, att TEXT, date INTEGER, FOREIGN KEY (uid) REFERENCES users(id))")
+            })
+        }
         this.db = db
     }
     addUser(uObj) {
@@ -22,5 +24,12 @@ export default class WebSQLAdapter {
     }
     q(txt) {
         return txt.replace(/'/g, '"')
+    }
+    async check() {
+        return new Promise((resolve, reject) =>{
+            this.db.transaction(tx => {
+                tx.executeSql("SELECT * FROM users", [], ()=>resolve(), ()=>reject())
+            })
+        }).catch(e=>e) || false
     }
 }
